Changes from inheritance to state pattern.

+++ Cloudents.Core/CommandHandler/Admin/DeleteDocumentCommandHandler.cs
@@ -1,9 +1,10 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using Cloudents.Core.Command.Admin;
+﻿using Cloudents.Core.Command.Admin;
 using Cloudents.Core.Entities.Db;
+using Cloudents.Core.Enum;
 using Cloudents.Core.Event;
 using Cloudents.Core.Interfaces;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Cloudents.Core.CommandHandler.Admin
 {
@@ -19,16 +20,15 @@ namespace Cloudents.Core.CommandHandler.Admin
         public async Task ExecuteAsync(DeleteDocumentCommand message, CancellationToken token)
         {
             var document = await _documentRepository.GetAsync(message.Id, token);
-            switch (document)
+            if (document.State == ItemState.Deleted)
             {
-                case null:
-                case DocumentDeleted _:
-                    return;
-                default:
-                    document.Events.Add(new DocumentDeletedEvent(document));
-                    await _documentRepository.DeleteAsync(document, token);
-                    break;
+                return;
             }
+
+            document.Events.Add(new DocumentDeletedEvent(document));
+            await _documentRepository.DeleteAsync(document, token);
+
+
         }
     }
 }
